+++ packages/abap-deploy-config-inquirer/src/prompts/validators.ts
@@ -37,6 +37,8 @@ import {
 } from '../types';
 import { AdaptationProjectType } from '@sap-ux/axios-extension';
 import { AbapServiceProviderManager } from '../service-provider-utils/abap-service-provider';
+import type { IValidationLink } from '@sap-devx/yeoman-ui-types';
+import { ERROR_TYPE, ErrorHandler } from '@sap-ux/inquirer-common';
 
 const allowedPackagePrefixes = ['$', 'Z', 'Y', 'SAP'];
 
@@ -372,17 +374,25 @@ export async function validatePackageChoiceInput(
     input: PackageInputChoices,
     systemConfig: SystemConfig,
     backendTarget?: BackendTarget
-): Promise<boolean | string> {
+): Promise<boolean | string | IValidationLink> {
     if (input === PackageInputChoices.ListExistingChoice) {
-        const retrievedPackageList = await queryPackages('', systemConfig, backendTarget);
-        if (retrievedPackageList && retrievedPackageList.length > 0) {
-            return true;
-        } else {
-            return t('warnings.packageNotFound');
+        let helpLink: IValidationLink | string | undefined;
+        try {
+            const retrievedPackageList = await queryPackages('', systemConfig, backendTarget);
+            if (retrievedPackageList && retrievedPackageList.length > 0) {
+                return true;
+            } else {
+                return t('warnings.packageNotFound');
+            }
+        } catch (error) {
+            if (ErrorHandler.isCertError(error)) {
+                helpLink = ErrorHandler.getHelpForError(ERROR_TYPE.CERT);
+                return helpLink ?? true;
+            }
+            throw error;
         }
-    } else {
-        return true;
     }
+    return true;
 }
 
 /**
@@ -571,37 +581,38 @@ export async function validateTransportChoiceInput({
     backendTarget?: BackendTarget;
     ui5AbapRepoName?: string;
     transportDescription?: string;
-}): Promise<boolean | string> {
+}): Promise<boolean | string | IValidationLink> {
     const packageAnswer = getPackageAnswer(previousAnswers, PromptState.abapDeployConfig.package);
     const systemConfig = getSystemConfig(useStandalone, PromptState.abapDeployConfig, backendTarget);
 
-    switch (input) {
-        case TransportChoices.ListExistingChoice: {
-            return handleListExistingTransportChoice(
+    if (input === TransportChoices.ListExistingChoice) {
+        try {
+            return await handleListExistingTransportChoice(
                 packageAnswer,
                 systemConfig,
                 previousAnswers,
                 backendTarget,
                 ui5AbapRepoName
             );
+        } catch (error) {
+            if (ErrorHandler.isCertError(error)) {
+                return ErrorHandler.getHelpForError(ERROR_TYPE.CERT) ?? true;
+            }
         }
-        case TransportChoices.CreateNewChoice: {
-            return handleCreateNewTransportChoice({
-                packageAnswer,
-                systemConfig,
-                input,
-                previousAnswers,
-                validateInputChanged,
-                prevTransportInputChoice,
-                backendTarget,
-                ui5AbapRepoName,
-                transportDescription
-            });
-        }
-        case TransportChoices.EnterManualChoice:
-        default:
-            return true;
+    } else if (input === TransportChoices.CreateNewChoice) {
+        return await handleCreateNewTransportChoice({
+            packageAnswer,
+            systemConfig,
+            input,
+            previousAnswers,
+            validateInputChanged,
+            prevTransportInputChoice,
+            backendTarget,
+            ui5AbapRepoName,
+            transportDescription
+        });
     }
+    return true;
 }
 
 /**
@@ -708,8 +719,14 @@ export async function validatePackage(
         // we need to verify cloud systems are connected before checking the package to avoid multiple browser windows opening
         (PromptState.abapDeployConfig.scp && AbapServiceProviderManager.isConnected())
     ) {
-        // checks if package is a local package and will update prompt state accordingly
-        await getTransportListFromService(input.toUpperCase(), answers.ui5AbapRepo ?? '', backendTarget);
+        try {
+            // checks if package is a local package and will update prompt state accordingly
+            await getTransportListFromService(input.toUpperCase(), answers.ui5AbapRepo ?? '', backendTarget);
+        } catch (error) {
+            LoggerHelper.logger.warn(
+                `An error occurred while validating the local package for package: ${error.message}`
+            );
+        }
     }
 
     const startingPrefixValidation = validatePackageStartingPrefix(input, answers, promptOption, ui5AbapPromptOptions);