+++ examples/odata-cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/odata-cli
 
+## 0.17.4
+
+### Patch Changes
+
+-   Updated dependencies [b45093b]
+    -   @sap-ux/axios-extension@1.21.4
+
+## 0.17.3
+
+### Patch Changes
+
+-   Updated dependencies [4303f99]
+    -   @sap-ux/axios-extension@1.21.3
+
 ## 0.17.2
 
 ### Patch Changes