+++ packages/app-config-writer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/app-config-writer
 
+## 0.6.10
+
+### Patch Changes
+
+-   Updated dependencies [b45093b]
+    -   @sap-ux/axios-extension@1.21.4
+
+## 0.6.9
+
+### Patch Changes
+
+-   Updated dependencies [4303f99]
+    -   @sap-ux/axios-extension@1.21.3
+
 ## 0.6.8
 
 ### Patch Changes